user details WIP
