+++ api/collaboration/serializers.py
@@ -4,23 +4,9 @@ from rest_framework import serializers
 
 from api.collaboration.models import TeamMember
 from api.user.models import Profile
+from api.user.serializers import UserSerializer
 
 UserModel = get_user_model()
-
-
-class UserProfileSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Profile
-        fields = ["sso_user_id"]
-
-
-class UserSerializer(serializers.ModelSerializer):
-    profile = UserProfileSerializer()
-
-    class Meta:
-        model = UserModel
-        fields = ['profile', 'email', 'first_name', 'last_name']
 
 
 class BarrierTeamSerializer(serializers.ModelSerializer):
