+++ config/urls.py
@@ -22,7 +22,7 @@ from api.barriers.views import (
     BarrierStatusHistory,
 )
 from api.interactions.views import BarrierInteractionList, BarrierIneractionDetail
-from api.user.views import who_am_i
+from api.user.views import who_am_i, UserDetail
 from api.core.views import admin_override
 
 from api.interactions.urls import urlpatterns as interaction_urls
@@ -35,6 +35,7 @@ urlpatterns = [
     path("counts", barrier_count, name="barrier-count"),
     path("ping.xml", ping, name="ping"),
     path("whoami", who_am_i, name="who_am_i"),
+    path("users/<uuid:sso_user_id", UserDetail.as_view(), name="get-user"),
     path("reports", BarrierReportList.as_view(), name="list-reports"),
     path("reports/<uuid:pk>", BarrierReportDetail.as_view(), name="get-report"),
     path(
